+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Fix `Parser` failing on `template<... operator<...` functions ([pull #829](https://github.com/bytedeco/javacpp/pull/829))
  * Compile classes with `parameters` bumping minimum requirements to Java SE 8 and Android 7.0 ([issue bytedeco/javacpp-presets#1739](https://github.com/bytedeco/javacpp-presets/issues/1739))
 
 ### February 22, 2026 version 1.5.13